Folder division and code cleaning
